+++ app/Main.hs
@@ -64,6 +64,7 @@ data Options w = Options
   , attribute
       :: w ::: [Regex] <?> "Pattern (POSIX regex) to limit updates to expressions under matching names in attrsets and let bindings. Combined using inclusing or, if this isn't specified then no expressions will be filtered by attribute name"
   , dryRun :: w ::: Bool <!> "False" <?> "Don't modify the file"
+  , onlyCommented :: w ::: Bool <!> "False" <?> "Only update from Git sources which have a comment on the 'rev' (or 'url' for fetchTarball from GitHub) attribute"
   }
   deriving stock Generic
 
@@ -85,8 +86,9 @@ optParser =
                                       "attribute" -> Just 'A'
                                       n           -> firstLetter n
               , fieldNameModifier = \case
-                                      "dryRun" -> "dry-run"
-                                      n        -> n
+                                      "dryRun"        -> "dry-run"
+                                      "onlyCommented" -> "only-commented"
+                                      n               -> n
               }
             )
         <*> many