+++ tests/Samples.hs
@@ -62,7 +62,7 @@ runTest f =
         (System.Process.shell ("nix-store --init"))
         mempty
 
-  let env = Env (const (Data.Text.IO.hPutStrLn System.IO.stderr)) [] [] Wet
+  let env = Env (const (Data.Text.IO.hPutStrLn System.IO.stderr)) [] [] Wet False
   Update.Nix.FetchGit.processFile env (dir </> inBase)
 
   replaceFile (dir </> inBase) (Data.Text.pack dir) "/tmp/nix-update-fetchgit-test"