+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Sort tags with git ls-remote with `-v:refname` (according to version)
   - this allows one to have a comment like `# tags/v*` to get the latest
     version
+- Add --only-commented option to allow being a bit more explicit about what's
+  updated
 
 ## [0.2.3] - 2020-11-06